at-spi2-core: add dbus option.
Turning off the dbus option makes this into more or less the same as
the old atk library: that is, if you didn't have at-spi2-core
installed before, only atk, you can turn the option off to avoid
incurring a mandatory dbus dependency in all sorts of things that
didn't previously have one, like gtk2.

If you did have at-spi2-core installed before because things actually
used it, turning off the dbus option off will break those things. (As
in, they won't even build.)

It would be better if turning off dbus provided the at-spi2-core
libraries in a way that didn't depend on dbus, so things that wanted
them could build, but that's not the hook that upstream gives us.
(It's also not clear from very casual inspection whether at-spi2 is
architected adequately to be transport-independent.)

It also might be better to go back to having a separate package for
the atk library, but this is also not so easy to arrange.

Therefore, the dbus option is enabled by default here.