sysutils/xentools418: Fix build when pandoc is installed
xentools418 does not TOOLS_DEPEND on pandoc, but it looks for it and
finds an installed copy.  This breaks the build because of some latex
issue, and would if it succeeded likely install more files, changing
the binary package.  Therefore, treat usage of a
non-manifestly-depended-on tool as a bug and patch it out.  The patch
is kludgy because configure.ac doesn't support the normal approaches.

With this, package builds on NetBSD 10 amd64 with pandoc present.