x11/ocaml-lablgtk3: Kludge around dune pkg-config issues
For unclear reasons, if PKG_CONFIG is set in the environment, even if
properly to a copy of pkg-config in the work directory, dune objects
to it and errors out.  Remove PKG_CONFIG from the build environment.
This allows building to succeed.

This commit adds some test targets, not invoked, with comments that
are notes to the next person struggling with this, perhaps future
self.

The removal of PKG_CONFIG is implemented by open-coding the do-build
target from ocaml.mk, pending disscussion of the wisdom of changing
that target, on the theory that this is a dune bug, not a lablgtk bug.