7-zip: updated to 24.08
7-Zip 24.08

The bug in 7-Zip 24.00-24.07 was fixed:
For creating a zip archive: 7-Zip could write extra zero bytes after the end of the archive, if a file included to archive cannot be compressed to a size smaller than original.
The created zip archive is correct except for the useless zero bytes after the end of the archive.
When unpacking such a zip archive, 7-Zip displays a warning:
"WARNING: There are data after the end of archive".
The bug was fixed: there was a leak of GDI objects (internal resources in Windows) in "Confirm File Replace" window, causing problems after 1600 displays of "Confirm File Replace" window from same running 7-Zip process.
Some optimizations for displaying file icons in 7-Zip File Manager and in "Confirm File Replace" window.
Some bugs were fixed.