math/py-tables: Hack to fix PLIST issue on macOS
setup.py fails to set the runtime dir properly, leading to
the mistaken impression that blosc libraries were not found,
even though it correctly identified headers and libs.
Added a heavily documented temporary patch to override this
error.  Upstream is aware of the issue, but I suspect may
take a while to get a handle on it.
https://github.com/PyTables/PyTables/issues/1219