wm/fvwm3: switch build system to meson.
fvwm3 migration to meson [0] marks te deprecation of the historical
infrastucture based on autotools. Apparently, switching to a different
and more flexible framework was easier than patching for autoconf 2.70.
As of 2024-09-30, there is a six-month timer on autools support in
fvwm3 [1].

In addition, we turn the go option on by default, since it's required by
FvwmPrompt, and will probably be required by any new Fvwm module.

Revision bumped.

[0] https://github.com/fvwmorg/fvwm3/discussions/1068
[1] https://mastodon.social/@thomasadam@bsd.network/113573995984463278