mit-krb5: Stop disabling thread support on SunOS.
This was blanket added in 2005 with no reason provided.  It is actively
harmful now on SunOS when trying to use threaded clients such as pgbench.
For now I'm only removing it on SunOS as that's all I can test, but it is
highly likely this should just be removed completely.

Bump PKGREVISION.