py-hg-evolve: update to 11.1.6.
11.1.6

This is a bugfix release. The notable changes are: evolve won't consider
all parts of a split content-divergent if only some of them are;
obshashrange cache will simply reset if there are too many affected
ranges instead of precision-deleting them from the database, this
improves overall performance; topic stack now provides any suitable
successor for an obsolete stack base instead of raising an error, this
prevents errors in commands like `hg topics --age` in certain cases.

11.1.5

This is a bugfix release. The most notable change is the compatibility
with the upcoming Mercurial 6.9. Also the cleanup step of `hg evolve
--abort` tries to be more careful when accessing revisions that might
not exist. Documentation and tests were updated to be closer to the
modern Python packaging standards.