geography/gama: Stop using ancient vendored expat
gama has a vendored ancient expat (to accomodate users of deficient
operating systems).  The build defaults to using it; this is of course
a bug and has been reported upstream.  This commit makes two changes:
  - in post-extract, remove the old expat sources to guard against use
  - pass a configure flag to look for expat normally

Adjust comments about gmake (upstream documents that now) and a
missing test depends (not changed due to freeze).

Builds and passes tests on NetBSD 10 amd64.  Likely works elsewhere,
but this is a security bug and it's better to have it fail and fix
with a pullup than leave it.  (gama is a leaf package.)