databases/ruby-tiny_tds: update to 3.0.0
3.0.0 (2024-12-15)

For a major release, there are no changes to the API of tiny_tds, and we
assume most users can upgrade immediately.

We dropped support for older Ruby and Microsoft SQL server versions.  We
also updated all dependencies for the precompiled Windows gem versions,
mainly OpenSSL v1 to v3.  We no longer provide a precompiled version for
32-bit versions of Windows.  And, if tiny_tds cannot send your SQL command
to the server, it will now raise an error message instead of returning
false.

Detailed changelog:

* Drop support for Ruby < 2.7
* Drop support for SQL Server < 2017
* Drop support for FreeTDS < 1.0
* No longer provide a 32-bit Windows build
* Raise error if FreeTDS is unable to send command buffer to the server
* Use freetds v1.4.23, libiconv v1.17 and OpenSSL v3.4.0 for Windows builds
* Add bigdecimal to dependencies