devel/ruby-activemodel80: add package version 8.0.1
Active Model -- model interfaces for Rails

Active Model provides a known set of interfaces for usage in model classes.
They allow for Action Pack helpers to interact with non-Active Record
models, for example.  Active Model also helps with building custom ORMs for
use outside of the Rails framework.

This is for Ruby on Rails 8.0.