xfce4-dev-tools: add meson as tool dependency
This looks unorthodox, because meson isn't actually required to build
this package; rather, the configure script expects it to be present in
the environment as a validation, and will fail without it. Since other
Xfce packages will require it anyway, seems simpler than editing the
configure script so this builds in sandboxed environments.