editors/tp-note: update to 1.25.0
New feature: detect the first heading in the HTML clipboard

A new filter `html_heading` used in the `from_clipboard_content` template
extracts the first HTML heading in the HTML clipboard. When you copy a chapter
with a heading in your browser, Tp-Note will place this heading in the `title:`
field of the new note's YAML header.