caddy: update to 2.9.1
2.9.1
-----

his is a patch release fixing a few issues related to FastCGI, HTTP/3
and unix sockets, log file permissions, repeated TERM signals, and an
ARI edge case. It also adds Caddyfile support for response matching in
the header directive.

2.9.0
-----

Aside from some minor new features, this release is mainly focused on
refinements and bug fixes in many areas, including:

- Config loading
- Events
- Logging
- Placeholders
- Reverse proxy and HTTP server performance
- Matchers
- HTTP (esp. HTTP/3)
- Metrics (per-host metrics)
- Security
- TLS automation & ACME ARI

We realize there is extensive interest in Encrypted Client Hello (ECH)
and post-quantum ciphers. These are slated to be supported in Go 1.24,
which is scheduled for a stable release in approximately February. We
did not want to force users to go through the inconvenience of
installing pre-release, non-stock installations of Go, even though the
RCs are quite stable and production-ready, in order to even compile
Caddy, which is quite common given our plugin ecosystem. We anticipate a
Caddy 2.10 release in the near future with these capabilities, built on
Go 1.24.