math/R: fix rpath flag in Makeconf
The R package installs a Makefile fragment, Makeconf, which includes
all the variable definitions required for R packages to compile
external programs.  The variable LIBR0, defined by the configure
script, is inserted into Makeconf and contains the linker directory
search and rpath arguments.  However, this variable passes the rpath
directory to the linker with -Wl,-R, which fails on Darwin.  Instead,
use ${COMPILER_RPATH_FLAG}, which is set correctly for each platform.

While here, make the paths for -rpath and -L the same.