devel/py-boost: Install cmake files needed by downstream packages.
This uses an intermediate package devel/py-boost-cmake which is
agnostic to the Python version, because boost/cmake now expects to
find unversioned files boost_python-config.cmake and
boost_python-config-version.cmake in canonical locations in order to
find Python-versioned stuff.

It's all kind of a mess -- and the py-boost-cmake package does a lot
of redundant work to build boost python libraries that it doesn't
install.  But it's currently only used by finance/ledger (and
finance/ledger32), so this mess doesn't affect much and if you want
to clean it up there's not much fallout to worry about.